check

+++ credit_model/config/core.py
@@ -47,23 +47,23 @@ class ModelConfig(BaseModel):
     unused_fields: List[str]
     
     checking_status: str
-    duration: float
+    duration: int
     credit_history: str
     purpose: str
-    credit_amount: float
+    credit_amount: int
     savings_status: str
     employment: str
-    installment_commitment: float
+    installment_commitment: int
     personal_status: str
     other_parties: str
-    residence_since: float
+    residence_since: int
     property_magnitude: str
-    age: float
+    age: int
     other_payment_plans: str
     housing: str
-    existing_credits: float
+    existing_credits: int
     job: str
-    num_dependents: float
+    num_dependents: int
         
     checking_mapping: Dict[str, int]
     credit_mappings: Dict[str, int]
